Patch: Don't manually add an extension when saving a file
When saving a file, the user can choose not to type a file extension.

In such cases, we can't just manually add a file extension later and
expect that writing to this new path will just work, because the
FileChooser portal did not allow us to write to this modified path.