Add support for NVENC encoding
This was tested on GTX 660 with the 470.239.06 driver, and only H.264
encoding was tested.

The version of the ffnvcodec module is fixed to use the 11.x branch on
purpose, so that older NVIDIA GPUs (like the GTX 660) can still benefit
from NVENC encoding as well.

Finally, a patch had to be added so that Avidemux could find the
ffnvcodec headers.