Re-add --filesystem=host permission
This permission was removed by me in commit 8771a298, but I'm re-adding
it now, because when you open an MPEG-TS file (.ts), Avidemux
immediately creates an index file (.idx2), and then tries to save it
next to the file you've opened.

For example:

1. Open a .ts file with Avidemux, e.g.: ~/video.ts
2. Avidemux will process the video file, and will immediately create an
   index file here: ~/video.ts.idx2

Without the --filesystem=host permission, i.e. with restricted
filesystem access, that index file is still created by the xdg-document
portal, but with a random name and also hidden, e.g.:

~/.xdp-video.ts.idx2-orEKzA

Despite of that, Avidemux still seems able to handle .ts files just
fine, though? In my limited testing, that is...

However, that hidden, random .idx2 file is never removed, for some
reason. And if we re-open that same .ts video, another hidden file will
be created.

So, to avoid these lingering .idx2 files, let's allow Avidemux
unrestricted access to the filesystem, until a proper fix or workaround
is found.

With this change, I can also remove the patch I added in 7e4cf947.